+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Implemented enhancements
 - XOT-856 - Staff SSO Integration for Admin Centre including 401 error if user not given admin rights
+- XOT-1077 - Get case studies from CMS on homepage
 
 
 ## [2019.09.23] (https://github.com/uktrade/navigator/releases/tag/2019.09.23)